+++ workers/ohsome_quality_analyst/api/api.py
@@ -1,14 +1,15 @@
 import fnmatch
 import json
 import logging
+from typing import Annotated
 
-from fastapi import Body, FastAPI, Request, status
+from fastapi import Body, FastAPI, Path, Request, status
 from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from geojson import Feature, FeatureCollection
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 
 from ohsome_quality_analyst import (
     __author__,
@@ -26,28 +27,27 @@ from ohsome_quality_analyst.api.request_models import (
     IndicatorData,
     IndicatorDatabase,
     IndicatorEnum,
+    ProjectEnum,
     ReportBpolys,
     ReportDatabase,
+    ReportEnum,
     TopicEnum,
 )
 from ohsome_quality_analyst.api.response_models import (
-    IndicatorMetadataListResponse,
     IndicatorMetadataResponse,
-    TopicListResponse,
-    TopicResponse,
+    MetadataResponse,
+    ReportMetadataResponse,
+    TopicMetadataResponse,
 )
 from ohsome_quality_analyst.config import configure_logging
 from ohsome_quality_analyst.definitions import (
     ATTRIBUTION_URL,
     get_attribution,
-    get_dataset_names,
-    get_fid_fields,
-    get_indicator_names,
+    get_indicator_definitions,
     get_metadata,
-    get_report_names,
+    get_report_definitions,
     get_topic_definition,
-    load_metadata,
-    load_topic_definitions,
+    get_topic_definitions,
 )
 from ohsome_quality_analyst.geodatabase import client as db_client
 from ohsome_quality_analyst.utils.exceptions import (
@@ -57,13 +57,15 @@ from ohsome_quality_analyst.utils.exceptions import (
     RasterDatasetUndefinedError,
     SizeRestrictionError,
     TopicDataSchemaError,
+    ValidationError,
 )
 from ohsome_quality_analyst.utils.helper import (
+    get_class_from_key,
     hyphen_to_camel,
     json_serialize,
-    name_to_class,
     snake_to_hyphen,
 )
+from ohsome_quality_analyst.utils.validators import validate_indicator_topic_combination
 
 MEDIA_TYPE_GEOJSON = "application/geo+json"
 
@@ -106,6 +108,9 @@ TAGS_METADATA = [
     },
 ]
 
+# TODO: to be replaced by config
+DEFAULT_PROJECT = ProjectEnum.core
+
 configure_logging()
 logging.info("Logging enabled")
 logging.debug("Debugging output enabled")
@@ -136,20 +141,13 @@ class CustomJSONResponse(JSONResponse):
         return json.dumps(content, default=json_serialize).encode()
 
 
-@app.exception_handler(ValidationError)
 @app.exception_handler(RequestValidationError)
+@app.exception_handler(ValidationError)
 async def validation_exception_handler(
-    request: Request, exception: RequestValidationError | ValidationError
+    request: Request,
+    exception: RequestValidationError | ValidationError,
 ):
-    """Override request validation exceptions.
-
-    `pydantic` raises on exception regardless of the number of errors found.
-    The `ValidationError` will contain information about all the errors.
-
-    FastAPIs `RequestValidationError` is a subclass of pydantic's `ValidationError`.
-    Because of the usage of `@pydantic.validate_arguments` decorator
-    `ValidationError` needs to be specified in this handler as well.
-    """
+    """Exception handler for validation exceptions."""
     return JSONResponse(
         status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
         content=jsonable_encoder(
@@ -179,7 +177,7 @@ async def oqt_exception_handler(
         | SizeRestrictionError
     ),
 ):
-    """Exception handler for custom OQT exceptions."""
+    """Exception handler for OQT exceptions."""
     return JSONResponse(
         status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
         content={
@@ -208,20 +206,15 @@ class MappingSaturationModel(BaseModel):
         alias_generator = snake_to_hyphen
 
 
-# TODO (Experimental): Make this endpoint general and remove `/indicator` endpoint below
-@app.post("/indicators/mapping-saturation", tags=["indicator"], include_in_schema=False)
-async def post_indicator_mapping_saturation(
-    parameters: MappingSaturationModel,
-) -> CustomJSONResponse:
-    """Request an Indicator for an AOI."""
-    parameters = IndicatorBpolys(
-        name="mapping-saturation", topic=parameters.topic_key, bpolys=parameters.bpolys
-    )
-    return await post_indicator(parameters)
-
-
-@app.post("/indicator", tags=["indicator"])
+@app.post("/indicators/{key}", tags=["indicator"])
 async def post_indicator(
+    key: Annotated[
+        IndicatorEnum,
+        Path(
+            title="Indicator Key",
+            example="mapping-saturation",
+        ),
+    ],
     parameters: IndicatorBpolys
     | IndicatorDatabase
     | IndicatorData = Body(
@@ -230,8 +223,11 @@ async def post_indicator(
     ),
 ) -> CustomJSONResponse:
     """Request an Indicator for an AOI defined by OQT or a custom AOI."""
+    if isinstance(parameters, (IndicatorBpolys, IndicatorDatabase)):
+        validate_indicator_topic_combination(key.value, parameters.topic_key.value)
     geojson_object = await oqt.create_indicator_as_geojson(
         parameters,
+        key=key.value,
         size_restriction=True,
     )
     if parameters.include_svg is False:
@@ -247,25 +243,33 @@ async def post_indicator(
             parameters.flatten,
         )
     response = empty_api_response()
-    response["attribution"]["text"] = name_to_class(
+    response["attribution"]["text"] = get_class_from_key(
         class_type="indicator",
-        name=parameters.name.value,
+        key=key.value,
     ).attribution()
     response.update(geojson_object)
     return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_GEOJSON)
 
 
-@app.post("/report", tags=["report"])
+@app.post("/reports/{key}", tags=["report"])
 async def post_report(
+    key: Annotated[
+        ReportEnum,
+        Path(
+            title="Report Key",
+            example="building-report",
+        ),
+    ],
     parameters: ReportBpolys
     | ReportDatabase = Body(
         ...,
         examples=REPORT_EXAMPLES,
-    )
+    ),
 ) -> CustomJSONResponse:
     """Request a Report for an AOI defined by OQT or a custom AOI."""
     geojson_object = await oqt.create_report_as_geojson(
         parameters,
+        key=key.value,
         size_restriction=True,
     )
     if parameters.include_html is False:
@@ -281,8 +285,9 @@ async def post_report(
             parameters.flatten,
         )
     response = empty_api_response()
-    response["attribution"]["text"] = name_to_class(
-        class_type="report", name=parameters.name.value
+    response["attribution"]["text"] = get_class_from_key(
+        class_type="report",
+        key=key.value,
     ).attribution()
     response.update(geojson_object)
     return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_GEOJSON)
@@ -304,61 +309,70 @@ async def get_available_regions(asGeoJSON: bool = False):
         return response
 
 
-@app.get("/indicators")
-async def indicator_names():
-    """Get names of available indicators."""
-    response = empty_api_response()
-    response["result"] = get_indicator_names()
-    return response
-
-
-@app.get("/datasets")
-async def dataset_names():
-    """Get names of available datasets."""
-    response = empty_api_response()
-    response["result"] = get_dataset_names()
-    return response
-
-
-@app.get("/reports")
-async def report_names():
-    """Get names of available reports."""
-    response = empty_api_response()
-    response["result"] = get_report_names()
-    return response
-
-
-@app.get("/fid-fields")
-async def list_fid_fields():
-    """List available fid fields for each dataset."""
-    response = empty_api_response()
-    response["result"] = get_fid_fields()
-    return response
-
-
-@app.get("/metadata/topics", tags=["metadata"])
-async def metadata_topic() -> TopicListResponse:
+@app.get(
+    "/metadata",
+    tags=["metadata"],
+    response_model_exclude={
+        "result": {
+            "topics": {k.value: {"key": True} for k in TopicEnum},
+            "indicators": {
+                k.value: {"label_description": True, "result_description": True}
+                for k in IndicatorEnum
+            },
+            "reports": {k.value: {"label_description": True} for k in ReportEnum},
+        }
+    },
+)
+async def metadata(project: ProjectEnum = DEFAULT_PROJECT) -> MetadataResponse:
     """Get topics."""
-    return TopicListResponse(result=list(load_topic_definitions().values()))
+    result = {
+        "topics": get_topic_definitions(project=project.value),
+        "indicators": get_indicator_definitions(project=project.value),
+        "reports": get_report_definitions(project=project.value),
+    }
+    return MetadataResponse(result=result)
 
 
-@app.get("/metadata/topics/{key}", tags=["metadata"])
-async def metadata_topic_by_key(key: TopicEnum) -> TopicResponse:
+@app.get(
+    "/metadata/topics",
+    tags=["metadata"],
+    response_model_exclude={
+        "result": {k.value: {"key": True} for k in TopicEnum},
+    },
+)
+async def metadata_topic(
+    project: ProjectEnum = DEFAULT_PROJECT,
+) -> TopicMetadataResponse:
+    """Get topics."""
+    return TopicMetadataResponse(result=get_topic_definitions(project=project.value))
+
+
+@app.get(
+    "/metadata/topics/{key}",
+    tags=["metadata"],
+    response_model_exclude={"result": {k.value: {"key": True} for k in TopicEnum}},
+)
+async def metadata_topic_by_key(key: TopicEnum) -> TopicMetadataResponse:
     """Get topic by key."""
-    return TopicResponse(result=get_topic_definition(key.value))
+    return TopicMetadataResponse(result={key.value: get_topic_definition(key.value)})
 
 
 @app.get(
     "/metadata/indicators",
     tags=["metadata"],
     response_model_exclude={
-        "result": {"__all__": {"label_description": True, "result_description": True}}
+        "result": {
+            k.value: {"label_description": True, "result_description": True}
+            for k in IndicatorEnum
+        },
     },
 )
-async def metadata_indicators() -> IndicatorMetadataListResponse:
+async def metadata_indicators(
+    project: ProjectEnum = DEFAULT_PROJECT,
+) -> IndicatorMetadataResponse:
     """Get metadata of all indicators."""
-    return IndicatorMetadataListResponse(
-        result=list(load_metadata("indicators").values())
+    return IndicatorMetadataResponse(
+        result=get_indicator_definitions(project=project.value)
     )
 
 
@@ -366,13 +380,44 @@ async def metadata_indicators() -> IndicatorMetadataListResponse:
     "/metadata/indicators/{key}",
     tags=["metadata"],
     response_model_exclude={
-        "result": {"label_description": True, "result_description": True}
+        "result": {
+            k.value: {"label_description": True, "result_description": True}
+            for k in IndicatorEnum
+        }
     },
 )
 async def metadata_indicators_by_key(key: IndicatorEnum) -> IndicatorMetadataResponse:
     """Get metadata of an indicator by key."""
     return IndicatorMetadataResponse(
-        result=get_metadata("indicators", hyphen_to_camel(key.value))
+        result={key.value: get_metadata("indicators", hyphen_to_camel(key.value))}
+    )
+
+
+@app.get(
+    "/metadata/reports",
+    tags=["metadata"],
+    response_model_exclude={
+        "result": {k.value: {"label_description": True} for k in ReportEnum}
+    },
+)
+async def metadata_reports(
+    project: ProjectEnum = DEFAULT_PROJECT,
+) -> ReportMetadataResponse:
+    """Get metadata of all indicators."""
+    return ReportMetadataResponse(result=get_report_definitions(project=project.value))
+
+
+@app.get(
+    "/metadata/reports/{key}",
+    tags=["metadata"],
+    response_model_exclude={
+        "result": {k.value: {"label_description": True} for k in ReportEnum}
+    },
+)
+async def metadata_reports_by_key(key: ReportEnum) -> ReportMetadataResponse:
+    """Get metadata of an indicator by key."""
+    return ReportMetadataResponse(
+        result={key.value: get_metadata("reports", hyphen_to_camel(key.value))}
     )