+++ workers/ohsome_quality_analyst/indicators/models.py
@@ -13,6 +13,7 @@ class IndicatorMetadata(BaseModel):
     description: str
     label_description: dict
     result_description: str
+    project: str
 
     class Config:
         alias_generator = snake_to_hyphen
@@ -39,13 +40,13 @@ class Result(BaseModel):
     """
 
     description: str
-    svg: str
     html: str
     timestamp_oqt: datetime = datetime.now(timezone.utc)  # UTC datetime object
     timestamp_osm: datetime | None = None
     value: float | None = None
     class_: Literal[1, 2, 3, 4, 5] | None = None
     figure: dict | None = None
+    svg: str | None = None
 
     @property
     def label(self) -> Literal["green", "yellow", "red", "undefined"]: