+++ workers/ohsome_quality_analyst/ohsome/client.py
@@ -37,6 +37,7 @@ async def _(
     ratio: Optional[bool] = False,
     group_by_boundary: Optional[bool] = False,
     count_latest_contributions: Optional[bool] = False,
+    contribution_type: Optional[str] = None,
 ) -> dict:
     """Query ohsome API with given Topic definition and arguments.
 
@@ -51,9 +52,11 @@ async def _(
         group_by_boundary: Group by boundary.
         count_latest_contributions:  Count of the latest contributions provided to the
             OSM data.
+        contribution_type: filters contributions by contribution type: ‘creation’,
+            ‘deletion’, ‘tagChange’, ‘geometryChange’ or a combination of them.
     """
     url = build_url(topic, ratio, group_by_boundary, count_latest_contributions)
-    data = build_data_dict(topic, bpolys, time, ratio)
+    data = build_data_dict(topic, bpolys, time, ratio, contribution_type)
     response = await query_ohsome_api(url, data)
     return validate_query_results(response, ratio, group_by_boundary)
 
@@ -156,6 +159,7 @@ def build_data_dict(
     bpolys: Union[Feature, FeatureCollection],
     time: Optional[str] = None,
     ratio: Optional[bool] = False,
+    contribution_type: Optional[str] = None,
 ) -> dict:
     """Build data dictionary for ohsome API query.
 
@@ -173,6 +177,8 @@ def build_data_dict(
         data["time"] = time
     if ratio:
         data["filter2"] = topic.ratio_filter
+    if contribution_type is not None:
+        data["contributionType"] = contribution_type
     return data